Split AuxiliaryAsyncProfiler into AsyncProfiler
That is to ease reuse of AsyncProfiler-based profiling for OpenJ9
